Refactor config (#12)
* changed .INI config files to YAML and setup Server

Config files are now saved as YAML, which is a nicer format.

Further implemented a parameter server, that automatically updates the
parameters, whenever a change occurs.

Additionally it will broadcast an even whenever a parameter is changed,
allowing to dynamically reconfigure the robot during runtime.

* fixed some random file shenanigans

* removed base path

It was causing to much of a havock for its benefit. For now just specify
bsolute paths to the config.

May be changed if necessary.

* small changes

to integrate the parameter server into NAO's environment.