Refactor core (#13)
* started refactoring core

* example on how qiChat improves the code

* moved loadJSON to utility function

* some big refactoring

The goal is to get rid of Game.py and move it into GameModule.py

Also we now have a nice way to integrate speech, so I am integrating
this into the game.
Essentially the game loop is split into two parts:
1) the setup where the robot decides to ask a question or make a guess
and do respective preperation
2) a callback from speech, that deals with the result of asking either
of the two

Further the format files will move from JSON to YAML

* continued refactoring

* finished first chunk of refactor

removed Game.py and ported all functions into GameModule.py

rewrote question / Answer / Animal to reflect that change

created smoother flow of the game

* added question logging

the robot now logs answers given to questions asked (as understood)

* remove unnecessary files

* refactored main

so that the parameter server is called in the ususal (new) manner

* changed to relative paths

in GameModule now using relative paths, so that it can be called
on multiple machines and from different folders

* added logging for speechModule

* added logging + minor bugfix

* made top level config relative

* removed unnecessary file

* refactored questions

now stored using YAML instead of JSON.
Its easier to read / write for humans and looks cleaner

* bugfix + animal refactor

fixed a bug with relative vs. absolute pathing

refactored animals to now use yaml

* fixed path bug

changed the way files are handled to lists of names.
This allows to handle paths platform independant

* added alternative way to say yes

* core can now load abs paths as well

* refactored animals and questions in core

core now also supports platform independant paths =)

* patched bug

Fixed a bug, which caused label propabilities to not be loaded

* fixed another bug

the propabuility didn't guard against the case when all propabilities
were 0.
In this case simply set the expectation to 1.0 for each value. There is
no information to gain from this label